move to memento directory
