+++ src/main/java/fivefishes/design/patterns/group/assignment/entities/Memento/ChangErMemento.java
@@ -1,4 +1,4 @@
-package fivefishes.design.patterns.group.assignment.entities;
+package fivefishes.design.patterns.group.assignment.entities.Memento;
 
 public class ChangErMemento {
     private String fashionType;
